requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0.76
PyYAML>=6.0.1
